requirements_rag.txt
# RAG Resume Q&A Bot Requirements

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Vector search and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
torch>=2.1.0

# Optional: Advanced reranking
# bge-reranker-base  # Uncomment for reranking support

# Data processing
numpy>=1.24.0
pandas>=2.1.0

# HTTP client for LLM integration
requests>=2.31.0
httpx>=0.25.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and utilities
structlog>=23.2.0
python-multipart>=0.0.6

# Optional: For YAML support in data loader
pyyaml>=6.0
